lock: add a show desktop option
Under certain circumstances (eg. Remmina remote desktop connections) some
windows can retain focus even from a screen locking program, making it
impossible to unlock a session without killing the lock program.

Provide an option to 'lock' to hide all windows prior to locking and
restoring the state of the desktop following an unlock.  This preserves
the previous behaviour of being able to get a blurred / pixelated
screenshot of the desktop while ensuring that nothing but the root window
has focus before 'i3lock' starts up.

This depends on 'wmctrl' and the window manager being able to recognize
the _NET_SHOWING_DESKTOP EWMH message.  If 'wmctrl' is not present on the
system or the current window manager does not support
_NET_SHOWING_DESKTOP, the previous behaviour is unchanged.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>